fix(x64): clear the last 10 core WW3D2 truncating casts, unblock GameEngine
Five distinct sites, three different outcomes -- the reusable lesson is in
telling them apart, not in a mechanical widen-everything pass:

- surfaceclass.cpp (2 sites, 4 error occurrences across g_ww3d2/z_ww3d2):
  widened `(unsigned int)lock_rect.pBits` to `(uintptr_t)lock_rect.pBits`.
  pBits is a live D3D-locked-surface address, runtime-only, never crosses a
  format boundary -- the real fix.

- dx8webbrowser.cpp: NOT widened. CreateBrowser's `parentwindow` parameter
  is `long` per BrowserEngine.idl, a fixed-width COM/oleautomation ABI
  contract we don't control. Widening the local HWND would not help --
  the call still narrows to `long` regardless. Replaced the
  now-illegal `reinterpret_cast<long>(hWnd)` with the explicit, legal
  `(long)(uintptr_t)hWnd` and added a TODO(x64-hwnd-truncation) comment
  stating the top 32 bits of the window handle are silently dropped on
  64-bit targets. Confirmed via -fsyntax-only that this path is live on
  MinGW (both targets take the non-_MSC_VER `#import` branch), not dead
  code.

- assetmgr.cpp (Generals + GeneralsMD, 2 casts each): not a width problem
  at all. `((int)mesh_name) - ((int)name) + 1` looked like it could be a
  hash or identity token (the brief's suspicion), but traced to plain
  pointer subtraction miscoded as two independent truncating casts before
  subtracting -- a real x64 bug (wrong length if the two 32-bit-truncated
  addresses straddle a wraparound differently), not cast noise. Fixed with
  correct pointer arithmetic: `(int)(mesh_name - name) + 1`, narrowing
  only the final result (a filename length) where lstrcpynA's Win32
  signature requires `int`.

None of the four touched files previously included BaseTypeCore.h. That
header carries `#pragma warning(error : 4706/4189/4101)` plus MIN/MAX/
TRUE/FALSE macros -- pulling it cold into files that never had it risks
promoting a pre-existing warning to a hard error under the VC6 G3 gate,
which is worse than the defect being fixed. Used the same narrower
`<Utility/stdint_adapter.h>` + raw `uintptr_t`/`intptr_t` approach the
earlier huffencode.cpp fix used for this identical situation -- same
types BaseTypeCore.h's UnsignedIntPtr/IntPtr alias, none of the pragma or
macro footprint.

Gates:
- G1 (32-bit control): exit 0, 0 errors, 31,555 warnings -- exact baseline.
- G3 (VC6 retail): exit 0, 0 errors, all seven .text digests
  byte-identical to docs/x64/baseline-vc6.md, despite two of the four
  files being renderer files the VC6 build compiles.
- G2 (x64): errors 10->0 as intended, but the real result is depth: this
  unblocked g_ww3d2/z_ww3d2 and with them GameEngine, which had never
  been measured as 64-bit before. Objects built 528->1,148 of 1,889
  (27.9%->60.8%), targets 42->44, GameEngine mentions in the log 0->13,105.
  New error count 10->602 (36 distinct shapes, 102 files) is entirely
  newly-visible GameEngine defects, not regression from this change.

+++ Core/Libraries/Source/WWVegas/WW3D2/surfaceclass.cpp
@@ -53,6 +53,9 @@
 #include "WWMath/vector2i.h"
 #include "colorspace.h"
 #include "WWLib/bound.h"
+// Only pulls the pointer-sized-int typedefs (uintptr_t); avoids dragging in
+// BaseTypeCore.h's warning-as-error pragmas into a file that never had them.
+#include <Utility/stdint_adapter.h>
 #include <d3dx8.h>
 
 void Convert_Pixel(Vector3 &rgb, const SurfaceClass::SurfaceDescription &sd, const unsigned char * pixel)
@@ -575,7 +578,10 @@ void SurfaceClass::FindBB(Vector2i *min,Vector2i*max)
 		for (x = min->I; x < max->I; x++) {
 
 			// HY - this is not endian safe
-			unsigned char *alpha=(unsigned char*) ((unsigned int)lock_rect.pBits+(y-min->J)*lock_rect.Pitch+(x-min->I)*size);
+			// pBits is a live D3D-locked surface pointer; the arithmetic below
+			// walks it row/column by row/column and stays runtime-only (never
+			// serialized), so widen the holder to pointer size on x64.
+			unsigned char *alpha=(unsigned char*) ((uintptr_t)lock_rect.pBits+(y-min->J)*lock_rect.Pitch+(x-min->I)*size);
 			unsigned char myalpha=alpha[size-1];
 			myalpha=(myalpha>>(8-alphabits)) & mask;
 			if (myalpha) {
@@ -649,7 +655,9 @@ bool SurfaceClass::Is_Transparent_Column(unsigned int column)
 	for (y = 0; y < (int) sd.Height; y++)
 	{
 		// HY - this is not endian safe
-		unsigned char *alpha=(unsigned char*) ((unsigned int)lock_rect.pBits+y*lock_rect.Pitch);
+		// Same live D3D-locked surface pointer as above: runtime-only, widen
+		// the holder to pointer size on x64.
+		unsigned char *alpha=(unsigned char*) ((uintptr_t)lock_rect.pBits+y*lock_rect.Pitch);
 		unsigned char myalpha=alpha[size-1];
 		myalpha=(myalpha>>(8-alphabits)) & mask;
 		if (myalpha) {
